Updated openstack/openstack Project: openstack/cinder  876f8927ce9a7e48a6f6b1d398ccf5b2495fcbcb
Fixes Windows Volume Driver upload volume fails

The copy_volume_to_image requires in some cases an intermediary
conversion, which takes place in the configured image conversion
dir. If this folder is not created, the operation will fail.

This patch ensures that method will create the folder if it does
not exist.

Change-Id: I82b5e76c6d6f3114666f184672667dc3b95bab8b
Closes-Bug: #1356415