Updated openstack/openstack Project: openstack/tempest  55810458b81a4b3d4e1445df4e07013ee2528357
Fix errors on import with service tags

When skips were added to service tags the CONF check was being
evaluated at import. This will cause an exception when there is
no config file. This commit fixes the issue by moving the config
eval inside the decorator wrapper so that it is only done at run
time.

Change-Id: I4127e2b837e103a8e0c24b324458bee97a2326ec