Updated openstack/openstack Project: openstack/keystone  5424e885a1ead181e8c0807b5ccc18c528dd348d
Add characterization test for cleanup role assignments for group

There was no test that showed that when a group is deleted the
role assignments for the group are expected to also be removed.
The test shows that the role assignments are not removed for the
LDAP and KVS assignment backends.

Change-Id: I7a11edcf516740fd2e73f0482ce6bbd67f5ab23f
Partial-Bug: #1366211