Updated openstack/openstack Project: openstack/horizon  4f1031184334cfa76937f92422b86587b5287326
Set the correct min_disk size when creating volume from image

Sometimes the image min_disk size is specified in a property rather
than the min_disk attribute (for instance, after creating a snapshot
from an instance). This ensures that the correct tooltip and error
message are displayed in these cases too.

Change-Id: I0c7a87a6f00c4cc1a54f8b70e5df9bb1d05de866
Closes-Bug: #1368600