Updated openstack/openstack Project: openstack/nova  809df65e7fbb234bda0a78ea47eb2fdf0a7f5a78
Add quotas for Server Groups (V2 API change)

This second change in a sequence that adds new quota values
for server groups.  This part adds the new V2 API extensions
to make the new quota values visible and changeable, and updates
the unit tests to check both V2 cases.

Note that Tempest changes https://review.openstack.org/#/c/112474/
is required to get a clean run as it checks for a specific
set of values.

The quota checks themselves are covered in a subsequent
dependent change.

Implements: blueprint server-group-quotas
DocImpact

Change-Id: I78974602d4be04deaf173b3e43f2dab92e8f4171