Updated openstack/openstack Project: openstack/nova  df7b6e7700b801e0c7f0dcbbcb08d8161da87b0d
neutronv2: attempt to delete all ports

Previously when deleting ports during allocate
or deallocate, the operation would fail on the
first non-404 error encountered. We now loop
through all of the ports and attempt the delete.

In the allocate flow we don't raise any errors
since the instance is set to ERROR anyway and
we're just trying to cleanup.

In the deallocate flow we attempt to delete all
of the ports and if there was a non-404 exception
we raise that at the end so the overall instance
delete operation fails.

Change-Id: Ibf68e0b4770b09ee32f187b85f2ca96984117e0b
Closes-bug: #1292243
Co-Authored-By: Dan Smith <dansmith@redhat.com>
Co-Authored-By: Sean Dague <sean@dague.net>
Co-Authored-By: Matt Riedemann <mriedem@us.ibm.com>