Updated openstack/openstack Project: openstack/nova  c79abb625c0a509733a4c0f12c1006d894d4a2d8
Stash numa-related flavor extra_spec items in system_metadata

This is a temporary measure to avoid having to add additional persistence
of extra_specs data simply to facilitate merging the NUMA patches. Only
the numa_cpus.* and numa_mem.* items are saved from extra_specs at
this time. The expectation is that this will quickly come out in kilo
when the rest of the flavor information stashed in system_metadata gets
removed and stored elsewhere.

Blueprint: virt-driver-numa-placement
Change-Id: I83ca863d521c579d8f1fe6168d7b1096af9f14f1