Updated openstack/openstack Project: openstack/nova  9fdddeb6d3ef399d08da21adc1f182f3ee9c50ea
API boot process sets NUMA topology for instances

This patch makes the API instance boot path look at the bits that can
contain requested topology, generate the desired topology and make sure
it is saved on the instance itself.

We also need to account for the flavor passed to
VirtNUMAInstanceTopology.get_constraints not being an actual Flavor
object, and work with db layer objects.

Change-Id: Ia3f3446da99a04e2d8152acb7d0bda7a1d502c83
Blueprint: virt-driver-numa-placement