Updated openstack/openstack Project: openstack/neutron  3b7e98801e076fa56eb3a62d7a172d83b1aa0367
Fix a test_db_plugin unit test side_effect usage

One of the unit tests didn't correctly use the side_effect
parameter for a mock to return different responses to multiple
calls, which resulted in an extra try-except block and an inline
note. This commit fixes the side_effect and removes the extra
catching code.

Change-Id: Ifa1a6a15025b710ecbba956b5b848e32314cf0cf