Updated openstack/openstack Project: openstack/neutron  696ff699475c956b116c46c59d45c245e9ee37db
Remove logic for conditional migrations

This patch removes the logic for handling conditional migrations
and adjusts all the patches where it was used accordingly.

The alembic environment has been update to not send anymore
the active plugins list to migrations.

This patch also removes the vestigial 'options' parameter which
was sent to every migration but always set to None and never
used by any migration.

Implements blueprint bp/reorganize-migrations

Change-Id: I7285e0276b262a9ea5d22c456a5a8cf34c461a0c