Updated openstack/openstack Project: openstack/tempest  9d7d28b66f9aa3fe1220f418cc2d13dd19984429
Fix tox coverage job

To run the tempest unit tests with coverage enabled the coverage
module needs to be installed. This requirement is properly listed in
test-requirements.txt, however the coverage job definition wasn't
installing coverage. This commit fixes this oversight by adding the
test-requirements file as a dependency for the job.

Change-Id: Ic443c6c4a3166e6b198cbe2af0a77ff2f0a38e83