Updated openstack/openstack Project: openstack/horizon  9041432629140e49be544d78d57df11efad26101
Corrected dash.can_access call to take context

horizon_main_nav makes a call to dash.can_access which should take
context but was being passed the request, this patch corrects that.
Unit test added to prevent regression, and fix added to RbacHorizonTests
to clean up changes to Cats and Dogs test dashboards, as it was
affecting any test that followed.

Change-Id: I46d9880ba3ac46c23a8253a4afcf2c169d6f4dc8
Closes-Bug: 1368673