Updated openstack/openstack Project: openstack/python-heatclient  f27a3571b7475620133cb0b314f4fb0d054f01f0
Add --show-nested option to stack-list

Adds an option to display nested stacks, provided the heat service
being called supports the show_nested argument.  In this mode we
display the parent stack ID as an additional column, a future TODO
is to add another switch which enables a tree/forest view as well,
which should be more friendly to human readers of the output.

Change-Id: Ic559d6c6467b1a2bc1058d792b99fea483cdfb65
blueprint: list-nested