Updated openstack/openstack Project: openstack/oslo.utils  e54a3597334f2cb10772950c97ddf2f6cc17b184
Enable mask_password to handle byte code strings

Byte code strings passed into mask_password cause the coersion to
six.text_type to fail with a UnicodeDecodeError.  The code can
work with the byte code string, but the incoming message may be
an i18n object that has to be coerced.

The temporary fix is to ignore the UnicodeDecodeError on the
coersion.   A better solution will be provided for Kilo.

This is related to cinder bug #1368527

Change-Id: Ifb60942b7cdfe4d184f59012c29caa9fa71fb053
Closes-bug: #1366189
Closes-bug: #1368527