Updated openstack/openstack Project: openstack/trove  d2d648f46d55424b92e83c326a0ae617b9b2f889
Fix NoSuchOptError on Couchbase create

A previous commit moved the backup_incremental_strategy option from
DEFAULT group to datastore specific grouping. However, the option was
only implemented for Mysql and Percona so when any datastore that had
backup support (like Couchbase) attempted to start, it received a
NoSuchOptError because of the missing option.

This fix put an empty dict in for the option for all datastores that
don't currently have backup or incremental backup support implemented.

Note: this wasn't picked up by any automated tests because the gate
currently only tests MySQL.

Author: Doug Shelley <doug@parelastic.com>
Co-Authored-By: Amrith Kumar <amrith@tesora.com>
Change-Id: I6a11dcf79c483a43ebdf79085d1bd17d653afe59
Closes-Bug: 1366175