Updated openstack/openstack Project: openstack-dev/pbr  eb7238725ed1837bd69d957f711087a45b82b8ab
Support reading versions from wheels

This change adds the ability for pbr to read the package version from
the METADATA file as well as the PKG-INFO file. The PKG-INFO file
is present for sdist packages but for wheel packages (PEP 427),
there is a METADATA file instead of a PKG-INFO file.

Closes-Bug: #1369179
Change-Id: I2fe52c846fd88d37a4ab78de87eec9defce1ba0a