Updated openstack/openstack Project: openstack/nova  6c36e9c411a9007def1c0748a64e2bb945ec3940
Fix race condition in update_dhcp

This modifies linux_net to only make one query for the fixed ips.
Previously it was getting the list of fixed ips twice. It also
merges the convoluted logic around single_default_gateway into
the single query so that it doesn't make a db request per fixed ip.
This prevents a race condition that could lead to ips not allocating
properly if an instance is deleted while the dhcp config was being
updated. Also, e274c45dcd37f49d66cc3cce72c2fd97f444151c missed a
call to obj_make_compatible for instance version. This adds it
since it will be in a conflicting spot.

Change-Id: Iffdc60063b26c9e3f6b8dbe5b67301cbc19b2564
Resolves-bug: #1359510