Updated openstack/openstack Project: openstack-dev/pbr  5edb427edc5c56f0af7ff72e28daf1afcc0c62a8
Refactor oneline changelog walking.

We want to suck the changelog into the description as well, but we
don't want all of it there. Making it a generator of formatted
objects lets us apply some logic without conflating things or tying
it up too much. This patch does prep work for the long term goal.

Change-Id: Idd00c219891500210c83faeb0f2d4370add3b069