Updated openstack/openstack Project: openstack/tempest  cc815761f2383486765f342b136e18e788287035
Update Heat CLI test to pass with random hash

The failing test assumes ordering of the returned JSON
blob and asserts that the given server resource precedes
another resource.  This updates it and allows the test
to pass when the resource is listed at the end of the
section, which may happen depending on hash + dict
ordering.

Change-Id: I279a1a1f3081214b0a2b0f712d311d148acd7b8e