Updated openstack/openstack Project: openstack/python-heatclient  3a05689c115a94927e813535b86c0c28665e0158
Reuse existing disable_rollback in stack-update

This patch makes `stack-update` command to change the `disable_rollback`
bit of the stack IF AND ONLY IF explicitly requested on CLI.

In order to facilitate this, `--rollback` CLI argument to `stack-update`
is added (accepting values from 1, on, true, t, 0, off, false, f and similar)
to explicitly set disable_rollback property of the stack on stack update.
If this argument is omitted, the value of disable_rollback
of the stack to be updated will be implicitly used by heat engine.

The boolean switch '--enable-rollback' is deprecated with appropriate
message (although the default behavior is effectively changed).

Change-Id: I235198ec04bfb192fcbfa4e2e3e618b64c99d3df
Closes-Bug: #1324366