Updated openstack/openstack Project: openstack/nova  207dcb53f5a62476716fd9976fd11a21a9518f01
libvirt: set pae for Xen PVM and HVM

The Physical Address Extension (PAE) enables paging to produce physical
addresses with more that 32 bit (2^32 bytes, 4 GBytes), namely it
extends the physical address space to 64 GBytes (2^36 bytes).

When spawning a 32 bit image in Xen PVM or HVM, PAE should be enabled in the
hypervisor so as to ensure that it is able to run properly. According to
the Xen documentation it can be enabled safely: "In general you should
leave this enabled and allow the guest Operating System to choose whether
or not to use PAE. " [1]

[1] http://xenbits.xen.org/docs/unstable/man/xl.cfg.5.html#pae_boolean

Closes-Bug: #1100697
Change-Id: If8d8ebc38efd969d96a1953446d72e72691734b3