Updated openstack/openstack Project: openstack/ceilometer  62575e463a13ebd634f237e510979731698faa2d
Switch to a custom NotImplementedError

Some drivers/modules might raise NotImplementedError whereas we would
expect things to work. Skipping test when such an error is raised would
be a bad idea, we would know the test failed.
So let's switch to a custom NotImplementedError and let's skip test only
if this one is raised.

Closes-Bug: #1369556

Change-Id: Ie88a281787218f9aba13b6e662eb4d49d3e685d0