Updated openstack/openstack Project: openstack/oslo.vmware  969bfba931f132199662851c6f309be67761c4f9
Switch to use requests/urllib3 and enable cacert validation

During Id: I422e6d4df4e392334c0ce47e6f5b017a8c95f61a we identified
rw_handles needed to be updated to requests and that we needed to
enable SSL certificate validation as well. This change consolidates
the create connection methods into FileHandle class and ensures that
additional headers and cacerts etc can be passed in for use during
the file upload/download process. Note that the create write connection
uses the httplib-equivalent support bundled in requests/urllib3 as
requests does not support the pattern of exposing a file like object
with multiple writes for the HTTP PUT operation, it only allows one
to pass a completely built file or file like object with data
already in memory. We need to be able to read buffers from somewhere
else (glance) and do a http put with that incremental data. This
usecase/pattern is not currently supported in requests.

Change-Id: I8a09a3fdf26bcaa3f4bdfbcccd22afbffd05af7c