Updated openstack/openstack Project: openstack/nova  00add127335a578009dcb322fa249b0d273ac589
Drop verbose and useless nova-api log information

Instead of printing two consecutive debug logs if there is a body, print
one. If no body, only display the method being called and not the
content-type as that isn't used for debugging. Shrink log output to make
log files smaller and easier to read.

Change-Id: I0e07876eee0064bd429ce9ea3254f7b95b991511