Updated openstack/openstack Project: openstack/nova  2cb9c85f8c2376b66e0de317dfc083998461010a
consolidate apirequest tests to single file

There was one stray apirequest test for a formatting check in
test_api. Bring it into the test_apirequest to for logical cohesion in
the code. This should also make it simpler for people to expand on
this validation later in one place if they want to.

Change-Id: Ib9a09034b1b677b462eb0384396cd9bc64f73045
Related-Bug: #1306479