Updated openstack/openstack Project: openstack/nova  10f36d78eccab0bd1778eb5dca3f3b55137bffce
Add a policy for handling retrospective vetos

For some time, we've been informally implementing a "revert early"
policy around handling -2 worthy complaints after a patch has been
merged. These reverts can be contentious, so it's appropriate we make
it very clear what the policy and procedure here is in order to lessen
the contentiousness of these reverts.

Co-authored-by: Michael Still <mikal@stillhq.com>

Change-Id: I90e4153a80c859a928630809aef31e7313f08906