Updated openstack/openstack Project: openstack/python-novaclient  0636d003c10bfbf7d15dc6a00762491aa640317d
novaclient: Convert v3 boot command with v2.1 spec (user-data)

The v3 client was written assuming that some arguments
would be renamed for better consistency.   V2.1 will follow
the v2 syntax,  so we need to convert the client back to
that.

This change strips the os-user-data: prefix from the user-data
field in the server element

Change-Id: I8f0e579b30547ce3fe778116745b4107d88c9be0