Updated openstack/openstack Project: openstack/python-openstackclient  d9a6f7fd3a26589f8d3cfb14fd568c93e9770c8e
Use Keystone client session.Session

This replaces the restapi requests wrapper with the one from Keystone client so
we can take advantage of the auth plugins.

As a first step only the v2 and v3 token and password plugins are supported.
This maintainis no changes to the command options or environment variables.

The next steps will include reworking the other API client interfaces to
fully utilize the single auth session.

Blueprint: ksc-session-auth
Change-Id: I47ec63291e4c3cf36c8061299a4764f60b36ab89