Updated openstack/openstack Project: openstack/heat  cd3e9103128aed44b7d85869d95c2f8314403bdf
Pass list of parameters to engine service to reset

For stack-update, a new argument named clear-parameters with
a list is passed in from the CLI and the new PATCH ReST API
to delete the indicated parameters from the existing parameters
in the DB, allowing the default parameters in the template to
be used.  A new method in environment handles the reset.

Partially-implements: blueprint troubleshooting-low-level-control
Partial-Bug: 1224828
Change-Id: Ia1270b679f27e264e6977c590d676b947c74c5da