Updated openstack/openstack Project: openstack/nova  eab855454be5e5056bc6442bd8cff88561e7ba82
Add more information to generic _add_floating_ip exception message

When we fail in some unexpected way in the _add_floating_ip method it'd
be good to have more information in the error message that's passed back
to the caller so they don't have to dig into the logs for the variables
involved, especially since the fixed IP might not be provided on the
request.

We're seeing messaging timeouts in the gate and it'd be nice to have the
error information right up front so we can fingerprint on it in
elastic-recheck.

Related-Bug: #1349590

Change-Id: I7c19c5c1bc5d41a922435d78aca0160c45130276