Updated openstack/openstack Project: openstack/nova  d0fc69096be9de94a40e65d5e24692b0ff788802
NUMA _get_constraint auto assumed Flavor object

We call this code in the API that is not converted to using objects at
the moment, so we cannot assume attribute access for dicts.

We also make sure that tests exercising it will fail by using dicts
exclusively until the code is converted to objects.

Change-Id: I6fb7104de5522c51f4706c973d8d8d8697d92b31
Closes-bug: #1369502