Updated openstack/openstack Project: openstack/heat  1fdce41373bd8d160915b2e23dd4b176bd332d49
Set the default state of a stack to CREATE_IN_PROGRESS

A stacks status is empty when originally entered into the database.
This leads to returning empty status responses back if the engine is
pre-occupied with other tasks.

This patch sets a stack's status to CREATE_IN_PROGRESS by default when
the stack object is created.

The stack object is only ever created during create and preview and is
loaded otherwise from the database in all the other scenarios.

Change-Id: Iab0892896b398530423bbe6c6a623376bf86590e
Closes-Bug: #1365399