Updated openstack/openstack Project: openstack/python-novaclient  3e5dae84a844f12cf27ea4800b96e18fb190148a
secgroup-create description is not optional

The description field to the Nova API for secgroup
create is not optional. This enforces the supplying
of the description field in novaclient as the Nova
call will fail if it is not supplied.

v3 version does not need to be updated as it was never
changed.

Change-Id: I6fc823990190b935a8e1214efc8c3ac3a0941abf
Closes-Bug: 1367121