Updated openstack/openstack Project: openstack/keystonemiddleware  ee477d461f41ea18889bdf4e907bc7e50d609898
Auth token tests create temp cert directory

The auth token tests were using the example signing directory as the
default signing directory. This is dangerous because the auth token
middleware may write to the directory in the course of the tests.
With this change, a temporary directory is created for each test
that's populated with the sample files.

Test this by doing `chmod -w examples/pki/certs` before running the
tests.

Change-Id: I139c6f0602d3258ac65ffe54bc0e7c3f82d14264