Updated openstack/openstack Project: openstack-dev/devstack  345420900574695b6b1c0f5fd7e7d618223b6864
The python-qpid packages must be installed when qpidd is configured

The python-qpid package is available for Ubuntu trusty, precise, and
all the supported RHEL based platforms.  This package is necessary if
qpidd is configured as the RPC backend.  It is the client API used to
talk to the broker, and must be installed on each system that
communicates with the broker.

Change-Id: I635d3e857aa4b769a80cb7cde405cfd6cae44d32