Updated openstack/openstack Project: openstack/tempest  f192f296cddb1a66f088628447feed7dde09a82c
Select AuthProvider type from credentials type

auth_version is used to get defaul credentials as well as to
select the AuthProvider class. If credentials of an API version
different from the configured one are used, a mismatch happens.

Instantiating the AuthProvider class that corresponds to the
version of API represented by the credentials object.

Closes-Bug: #1369557

Change-Id: I1ed6da90ce0abed805cf39190cb1ac3baa3f63b6