Updated openstack/openstack Project: openstack/nova  abb9db54107516eade3f868c133840940254f397
Fix NUMA fit testing in claims and filter class

This patch makes sure that the overall topology is considered when
checking wather a NUMA instance can fit on a NUMA host, as previously we
would only check the corresponding cells without considering the overall
topology.

It adds a method for checking it on the VirtNUMAHostTopology class and
make sure it's called when checking for placement in the scheduler and
confirming claims on compute hosts.

We also split out the method for getting the topology from a host
as it can be reused in the scheduler now.

Change-Id: I038c81a5241bf1fd2ce37f7eff69f89ecedf59ef
Closes-bug: #1369984