Updated openstack/openstack Project: openstack/tempest  c27feb8552dc9203f08992d8c6f0176538f0bf0a
Drop unused safe_setup decorator

The safe_setup decorator was introduced so that tests could
enforce tearDownClass being called in case of failures during
the setUpClass.

With the implementation of bp resource-cleanup, the test base
class automatically provides this facility for all tests,
so the safe_setup decorator is not needed anymore.

Partially-implements bp resource-cleanup

Change-Id: I3a4f34324420ee3a3ecc9a5e4e33d6ae27f64299