Updated openstack/openstack Project: openstack/horizon  9847358217c04c7e43bd71e8d241cca83adf0428
Workaround for negative vals in total*Used in nova absolute_limits

Due to bug 1370867 there is a case where negative values are returned
in total*Used in absolute_limits. If it occurs, Horizon Overview page
cannot be displayed due to an error "cannot convert float infinity to
integer". This commit replaces negative values in total*Used with 0
for a workarond.

Change-Id: I606803bdff1d335cc9f92d67619ff82f356a5143
Closes-Bug: #1370869